feat: email verification and forgot password
